requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pytest>=7.4.0
python-dotenv>=1.0.0
